// src/types/client.ts
export interface Client {
	id: number
	firstName?: string
	lastName?: string
	companyName?: string
	phone: string
	email?: string
	type: string
	address?: string
	codiceFiscale?: string
	partitaIVA?: string
	legalAddress?: string
	contactPerson?: string
	source?: string
	notes?: string
	createdAt?: string
	updatedAt?: string
}